+++ app/test/frontend/backend_test.dart
@@ -665,17 +665,13 @@ void main() {
         expect(version.sortOrder, 1);
 
         expect(versionPubspec.id, 'foobar_pkg-0.1.1+5');
-        expect(versionPubspec.namespace, '');
         expect(versionPubspec.package, testPackage.name);
-        expect(versionPubspec.qualifiedPackage, testPackage.name);
         expect(versionPubspec.version, testPackageVersion.version);
         expect(versionPubspec.updated.compareTo(dateBeforeTest) >= 0, isTrue);
         expect(versionPubspec.pubspec.asJson, loadYaml(testPackagePubspec));
 
         expect(versionInfo.id, 'foobar_pkg-0.1.1+5');
-        expect(versionInfo.namespace, '');
         expect(versionInfo.package, testPackage.name);
-        expect(versionInfo.qualifiedPackage, testPackage.name);
         expect(versionInfo.version, testPackageVersion.version);
         expect(versionInfo.updated.compareTo(dateBeforeTest) >= 0, isTrue);
         expect(versionInfo.readmeFilename, 'README.md');